correction.

+++ src/main/java/com/jingyunbank/etrade/user/entity/AddressEntity.java
@@ -9,19 +9,19 @@ public class AddressEntity extends BaseAddress{
 	 */
 	private static final long serialVersionUID = 1602851925715706080L;
 	
-	private int from = 0;
-	private int size = 10;
+	private long from = 0;
+	private long size = 10;
 	
-	public int getFrom() {
+	public long getFrom() {
 		return from;
 	}
-	public void setFrom(int from) {
+	public void setFrom(long from) {
 		this.from = from;
 	}
-	public int getSize() {
+	public long getSize() {
 		return size;
 	}
-	public void setSize(int size) {
+	public void setSize(long size) {
 		this.size = size;
 	}
